feat(framework): support Hailo raw image inputs
Add a Hailo-specific image classification loader, pass UINT8/NHWC runtime options through metadata, and make HailoRT handle explicit device IDs, scheduler activation without a context manager, batch size, and raw float-to-uint8 casting.

+++ framework/src/core/targets.py
@@ -362,7 +362,7 @@ register_target(TargetSpec(
     capabilities=("hef", "sync", "latency", "throughput", "monitor", "npu", "local"),
     runtime_options={
         "interface": "pcie",
-        "input_format_type": "float32",
+        "input_format_type": "uint8",
         "output_format_type": "uint8",
         "input_layout": "auto",
     },
